+++ src/components/SignIn/css/SocialLoginButton.module.css
@@ -4,7 +4,7 @@
   justify-content: center;
   width: 100%;
   max-width: 20rem;
-  padding: 0.5rem;
+  padding: 0.75rem 0.5rem;
   font-weight: bold;
   border: 1px solid var(--color-border-dark);
   border-radius: 0.5rem;
